requirements.txt
livekit-agents[google]
livekit-plugins-noise-cancellation
python-dotenv
requests
flask